sde: Add more details on how to extend the SDE
This patch adds more info about how to extend the SDE, how to add
more and different tests and adds the introduction page to the
sidebar of the blueprint.

It also removes the word PELUX from in front of all the mentions
of the SDE, this way it's easier to reuse the text in different
projects.