Add article about Private Docker Registry
Sometimes it's really convinient to run your own docker registry in
a project where you can host custom docker images which can be used
by the CI and the developers without the need to build them manually
every time you need them.

This article explains how to set up such a registry and push a
custom image to it.